LayoutTests/imported/w3c/web-platform-tests/css/css-page/page-rule-declarations-004-expected.txt


FAIL rule with invalid selector ['--a'] assert_not_equals: no @page rule should have been parsed got disallowed value 6
PASS adding a blank @page rule
PASS assigning invalid selector text ['1']
PASS assigning invalid selector text ['-3']
FAIL assigning invalid selector text ['--a'] assert_equals: should not be able to assign an invalid selector expected "" but got "--a"
FAIL assigning invalid selector text ['7cm'] assert_equals: should not be able to assign an invalid selector expected "" but got "--a"
FAIL assigning invalid selector text ['0.17'] assert_equals: should not be able to assign an invalid selector expected "" but got "--a"
FAIL assigning invalid selector text ['a, 123'] assert_equals: should not be able to assign an invalid selector expected "" but got "--a"
